added docker_running? and has_docker_compose conditions to nginx grid

+++ lib/smartcloud/grids/nginx.rb
@@ -7,16 +7,34 @@ module Smartcloud
 	
 			def self.start
 				puts "-----> Starting Nginx Network"
-				system("docker-compose -f #{self.docker_compose_filepath} up -d")
+				system("docker-compose -f #{self.docker_compose_filepath} up -d") if self.docker_running? && self.has_docker_compose?
 			end
 
 			def self.stop
 				puts "-----> Stopping Nginx Network"
-				system("docker-compose -f #{self.docker_compose_filepath} down")
+				system("docker-compose -f #{self.docker_compose_filepath} down") if self.docker_running? && self.has_docker_compose?
 			end
 	
 			def self.docker_compose_filepath
 				File.join(Smartcloud.root, 'lib/smartcloud/grids/grid-nginx/docker-compose.yml')
+			end
+
+			def self.docker_running?
+				if system("docker info")
+					true
+				else
+					puts "Error: Docker daemon is not running. Have you installed docker? Please ensure docker is running and try again."
+					false
+				end
+			end
+
+			def has_docker_compose?
+				if system("which docker-compose")
+					true
+				else
+					puts "Error: docker-compose is not installed. Please ensure docker-compose is running and try again."
+					false
+				end
 			end
 		end
 	end
